feat(fetchers): create api fetchers
