+++ app/open-source/page.jsx
@@ -1,5 +1,5 @@
 import { Container, Fade, Followers, GitHubProfile, RepositoryPreview, Title } from 'components';
-import { profileFetcher } from 'app/api/open-source/profile';
+import { profileFetcher } from 'lib/fetchers/open-source/fetcher';
 import meta from 'lib/config/metadata.js';
 
 export const metadata = {
